Add files via upload
I have gone through the code multiple times. I have had others go through it, I keep getting indent errors on it and cant not get it to run. I tried to put it in Jupyter and break down each section individually and I got all of it right but when i move it back to VS Code i get the same indent errors all over again. Fairly frustrating experience. I feel like this is a done project but again, i have not been able to run it due to errors from indentation.